[wdspec] Split fixture module for WebDriver HTTP and BiDi.
Differential Revision: https://phabricator.services.mozilla.com/D122336

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1694144
gecko-commit: 118ad1445b60ec5624ef4f5a125fc9716d5abe40
gecko-reviewers: webdriver-reviewers, jgraham
