+++ webdriver/tests/conftest.py
@@ -1,1 +1,5 @@
-pytest_plugins = "tests.support.fixtures"
+pytest_plugins = (
+    "tests.support.fixtures",
+    "tests.support.fixtures_bidi",
+    "tests.support.fixtures_http",
+)
