+++ webdriver/tests/support/fixtures.py
@@ -12,7 +12,6 @@ from tests.support import defaults
 from tests.support.helpers import cleanup_session, deep_update
 from tests.support.inline import build_inline
 from tests.support.http_request import HTTPRequest
-from tests.support.sync import Poll
 
 
 _current_session = None
@@ -21,8 +20,10 @@ _custom_session = False
 
 def pytest_configure(config):
     # register the capabilities marker
-    config.addinivalue_line("markers",
-        "capabilities: mark test to use capabilities")
+    config.addinivalue_line(
+        "markers",
+        "capabilities: mark test to use capabilities"
+    )
 
 
 @pytest.fixture
@@ -38,64 +39,15 @@ def pytest_generate_tests(metafunc):
             metafunc.parametrize("capabilities", marker.args, ids=None)
 
 
-# Ensure that the event loop is restarted once per session rather than the default  of once per test
-# if we don't do this, tests will try to reuse a closed event loop and fail with an error that the "future
-# belongs to a different loop"
+# Ensure that the event loop is restarted once per session rather than the default
+# of once per test. If we don't do this, tests will try to reuse a closed event
+# loop and fail with an error that the "future belongs to a different loop".
 @pytest.fixture(scope="session")
 def event_loop():
     """Change event_loop fixture to session level."""
     loop = asyncio.get_event_loop_policy().new_event_loop()
     yield loop
     loop.close()
-
-
-@pytest.fixture
-def add_event_listeners(session):
-    """Register listeners for tracked events on element."""
-    def add_event_listeners(element, tracked_events):
-        element.session.execute_script("""
-            let element = arguments[0];
-            let trackedEvents = arguments[1];
-
-            if (!("events" in window)) {
-              window.events = [];
-            }
-
-            for (var i = 0; i < trackedEvents.length; i++) {
-              element.addEventListener(trackedEvents[i], function (event) {
-                window.events.push(event.type);
-              });
-            }
-            """, args=(element, tracked_events))
-    return add_event_listeners
-
-
-@pytest.fixture
-def create_cookie(session, url):
-    """Create a cookie"""
-    def create_cookie(name, value, **kwargs):
-        if kwargs.get("path", None) is not None:
-            session.url = url(kwargs["path"])
-
-        session.set_cookie(name, value, **kwargs)
-        return session.cookies(name)
-
-    return create_cookie
-
-
-@pytest.fixture
-def create_frame(session):
-    """Create an `iframe` element in the current browsing context and insert it
-    into the document. Return a reference to the newly-created element."""
-    def create_frame():
-        append = """
-            var frame = document.createElement('iframe');
-            document.body.appendChild(frame);
-            return frame;
-        """
-        return session.execute_script(append)
-
-    return create_frame
 
 
 @pytest.fixture
@@ -144,7 +96,8 @@ async def session(capabilities, configuration, request):
     By default the session will stay open after each test, but we always try to start a
     new one and assume that if that fails there is already a valid session. This makes it
     possible to recover from some errors that might leave the session in a bad state, but
-    does not demand that we start a new session per test."""
+    does not demand that we start a new session per test.
+    """
     global _current_session
 
     # Update configuration capabilities with custom ones from the
@@ -179,12 +132,15 @@ async def session(capabilities, configuration, request):
 
 @pytest.fixture(scope="function")
 async def bidi_session(capabilities, configuration, request):
-    """Create and start a bidi session for a test that does not itself test
-    bidi session creation.
+    """Create and start a bidi session.
+
+    Can be used for a test that does not itself test bidi session creation.
+
     By default the session will stay open after each test, but we always try to start a
     new one and assume that if that fails there is already a valid session. This makes it
     possible to recover from some errors that might leave the session in a bad state, but
-    does not demand that we start a new session per test."""
+    does not demand that we start a new session per test.
+    """
     global _current_session
 
     # Update configuration capabilities with custom ones from the
@@ -236,93 +192,8 @@ def url(server_config):
 
 
 @pytest.fixture
-def create_dialog(session):
-    """Create a dialog (one of "alert", "prompt", or "confirm") and provide a
-    function to validate that the dialog has been "handled" (either accepted or
-    dismissed) by returning some value."""
-
-    def create_dialog(dialog_type, text=None):
-        assert dialog_type in ("alert", "confirm", "prompt"), (
-            "Invalid dialog type: '%s'" % dialog_type)
-
-        if text is None:
-            text = ""
-
-        assert isinstance(text, str), "`text` parameter must be a string"
-
-        # Script completes itself when the user prompt has been opened.
-        # For prompt() dialogs, add a value for the 'default' argument,
-        # as some user agents (IE, for example) do not produce consistent
-        # values for the default.
-        session.execute_async_script("""
-            let dialog_type = arguments[0];
-            let text = arguments[1];
-
-            setTimeout(function() {
-              if (dialog_type == 'prompt') {
-                window.dialog_return_value = window[dialog_type](text, '');
-              } else {
-                window.dialog_return_value = window[dialog_type](text);
-              }
-            }, 0);
-            """, args=(dialog_type, text))
-
-        wait = Poll(
-            session,
-            timeout=15,
-            ignored_exceptions=webdriver.NoSuchAlertException,
-            message="No user prompt with text '{}' detected".format(text))
-        wait.until(lambda s: s.alert.text == text)
-
-    return create_dialog
-
-
-@pytest.fixture
-def closed_frame(session, url):
-    original_handle = session.window_handle
-    new_handle = session.new_window()
-
-    session.window_handle = new_handle
-
-    session.url = url("/webdriver/tests/support/html/frames.html")
-
-    subframe = session.find.css("#sub-frame", all=False)
-    session.switch_frame(subframe)
-
-    deleteframe = session.find.css("#delete-frame", all=False)
-    session.switch_frame(deleteframe)
-
-    button = session.find.css("#remove-parent", all=False)
-    button.click()
-
-    yield
-
-    session.window.close()
-    assert new_handle not in session.handles, "Unable to close window {}".format(new_handle)
-
-    session.window_handle = original_handle
-
-
-@pytest.fixture
-def closed_window(session, inline):
-    original_handle = session.window_handle
-    new_handle = session.new_window()
-
-    session.window_handle = new_handle
-    session.url = inline("<input id='a' value='b'>")
-    element = session.find.css("input", all=False)
-
-    session.window.close()
-    assert new_handle not in session.handles, "Unable to close window {}".format(new_handle)
-
-    yield (original_handle, element)
-
-    session.window_handle = original_handle
-
-
-@pytest.fixture
 def inline(url):
-    """Takes a source extract and produces well-formed documents.
+    """Take a source extract and produces well-formed documents.
 
     Based on the desired document type, the extract is embedded with
     predefined boilerplate in order to produce well-formed documents.
@@ -350,7 +221,7 @@ def inline(url):
 
 @pytest.fixture
 def iframe(inline):
-    """Inlines document extract as the source document of an <iframe>."""
+    """Inline document extract as the source document of an <iframe>."""
     def iframe(src, **kwargs):
         return "<iframe src='{}'></iframe>".format(inline(src, **kwargs))
