+++ webdriver/tests/support/fixtures_bidi.py
@@ -1,0 +1,14 @@
+import asyncio
+from typing import Any, Mapping
+
+import pytest
+import webdriver
+
+
+@pytest.fixture
+def send_blocking_command(bidi_session):
+    """Send a blocking command that awaits until the BiDi response has been received."""
+    async def send_blocking_command(command: str, params: Mapping[str, Any]) -> Mapping[str, Any]:
+        future_response = await bidi_session.send_command(command, params)
+        return await future_response
+    return send_blocking_command
